net/nanocoap: add nanocoap_opt2 module
