+++ makefiles/pseudomodules.inc.mk
@@ -50,7 +50,8 @@ PSEUDOMODULES += lwip_tcp
 PSEUDOMODULES += lwip_udp
 PSEUDOMODULES += lwip_udplite
 PSEUDOMODULES += mpu_stack_guard
-PSEUDOMODULES += nanocoap_%
+PSEUDOMODULES += nanocoap_sock
+PSEUDOMODULES += nanocoap_opt2
 PSEUDOMODULES += netdev_default
 PSEUDOMODULES += netif
 PSEUDOMODULES += netstats
