+++ sys/net/application_layer/nanocoap/Makefile
@@ -1,3 +1,11 @@
 SRC := nanocoap.c
-SUBMODULES := 1
+
+ifneq (,$(filter nanocoap_sock,$(USEMODULE)))
+  SRC += sock.c
+endif
+
+ifneq (,$(filter nanocoap_opt2,$(USEMODULE)))
+  SRC += opt2_add.c
+endif
+
 include $(RIOTBASE)/Makefile.base
